requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9

